gateway/requirements.txt
# FastAPI Gateway Requirements
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# Async support
aiosqlite>=0.19.0

# HTTP client for webhook callbacks
httpx>=0.26.0
